add option to dock todo list window with Ctrl+D
This is to handle problems where the todo list won't show up at all when remoting into the lab machine if the todo list was undocked and on showing up on a secondary monitor. Ctrl+D toggles the todo list docking and it will revert to it's original state when undocked. This should be expanded to dock/undock any of the dock widgets but there is no easy way to do this with focus alone, maybe the context menu should be modified.